+++ gst-zed-src/gstzedsrc.h
@@ -99,21 +99,14 @@ struct _GstZedSrc {
     gint od_filter_mode;                             // sl::OBJECT_FILTERING_MODE filtering_mode
     gfloat od_prediction_timeout_s;                  // float prediction_timeout_s
     gboolean od_allow_reduced_precision_inference;   // bool allow_reduced_precision_inference
-    gfloat od_det_conf;                              // TODO(Walter) [runtime] float detection_confidence_threshold
-    gboolean od_person;                              // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_vehicle;                             // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_bag;                                 // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_animal;                              // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_electronics;                         // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_fruit_vegetable;                     // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_sport;                               // TODO(Walter) [runtime] std::vector< OBJECT_CLASS > object_class_filter
-    gboolean od_person_conf;                         // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
-    gboolean od_vehicle_conf;                        // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
-    gboolean od_bag_conf;                            // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
-    gboolean od_animal_conf;                         // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
-    gboolean od_electronics_conf;                    // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
-    gboolean od_fruit_vegetable_conf;                // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
-    gboolean od_sport_conf;                          // TODO(Walter) [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_det_conf;                              // [runtime] float detection_confidence_threshold
+    gfloat od_person_conf;                           // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_vehicle_conf;                          // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_bag_conf;                              // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_animal_conf;                           // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_electronics_conf;                      // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_fruit_vegetable_conf;                  // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
+    gfloat od_sport_conf;                            // [runtime] std::map< OBJECT_CLASS, float > object_class_detection_confidence_threshold
 
     gboolean body_tracking;           // TODO(Walter)
     gboolean bt_enable_segm_output;   // TODO(Walter) bool enable_segmentation